warning

+++ 2021/day19/day19.cpp
@@ -87,7 +87,7 @@ namespace day19 {
 
         for (const auto& rotation : all_3d_rotations) {
             std::copy(b.begin(), b.end(), b_rotated.begin());
-            for (int i = 0; i < b.size(); i++) {
+            for (size_t i = 0; i < b.size(); i++) {
                 ox::matrix<int>::in_place_multiplication(b_rotated[i], rotation, b[i]);
             }
